ai-services/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
aiofiles==0.7.0
pydantic==1.8.2
python-dateutil==2.8.2
regex==2023.12.25 